Add module param to bypassAll
