+++ ui/src/app/utils/access-utils.tsx
@@ -256,18 +256,8 @@ export const getAccessModuleBypassTime = (accessModules: Array<AccessModuleStatu
   return module ? module.bypassEpochMillis : null;
 };
 
-export const bypassAllModules = async(isBypassed: boolean) => {
-  // TypeScript enum iteration is nonfunctional
-  // so just copy the whole list
-  const modules = [
-    AccessModule.COMPLIANCETRAINING,
-    AccessModule.ERACOMMONS,
-    AccessModule.TWOFACTORAUTH,
-    AccessModule.DATAUSERCODEOFCONDUCT,
-    AccessModule.RASLINKLOGINGOV,
-  ];
-
-  for (const module of modules) {
+export const bypassAll = async(accessModules: AccessModule[], isBypassed: boolean) => {
+  for (const module of accessModules) {
     await profileApi().unsafeSelfBypassAccessRequirement({
       moduleName: module,
       isBypassed: isBypassed
